refactor types and schemas
