+++ apps/server/src/modules/memory/memory.types.ts
