+++ apps/server/src/modules/template/template.schema.ts
