+++ packages/shared/src/index.ts
@@ -1,32 +1,30 @@
 // Export all schema files
-export * from './schemas/integration.schema'
-export * from './schemas/canvas.schema'
+export * from '../../../apps/server/src/modules/integration/integration.schema'
 export * from './schemas/cli.schema'
-export * from './schemas/profile.schema'
-export * from './schemas/runtime.schema'
-export * from './schemas/platform.schema'
-export * from './schemas/billing.schema'
+export * from '../../../apps/server/src/modules/profile/profile.schema'
+export * from '../../../apps/server/src/modules/runtime/runtime.schema'
+export * from '../../../apps/server/src/modules/platform/platform.schema'
+export * from '../../../apps/server/src/modules/billing/billing.schema'
 export * from './schemas/generation.schema'
-export * from './schemas/template.schema'
-export * from './schemas/upload.schema'
-export * from './schemas/notification.schema'
-export * from './schemas/project.schema'
-export * from './schemas/usage.schema'
+export * from '../../../apps/server/src/modules/template/template.schema'
+export * from '../../../apps/server/src/modules/import/import.schema'
+export * from '../../../apps/server/src/modules/notification/notification.schema'
+export * from '../../../apps/server/src/modules/project/project.schema'
+export * from '../../../apps/server/src/modules/usage/usage.schema'
 
 // Export all types files
-export * from './types/integration.types'
-export * from './types/canvas.types'
+export * from '../../../apps/server/src/modules/integration/integration.types'
 export * from './types/cli.types'
 export * from './types/agent.types'
-export * from './types/memory.types'
+export * from '../../../apps/server/src/modules/memory/memory.types'
 export * from './types/profile.types'
 export * from './types/runtime.types'
-export * from './types/platform.types'
-export * from './types/billing.types'
+export * from '../../../apps/server/src/modules/platform/platform.types'
+export * from '../../../apps/server/src/modules/billing/billing.types'
 export * from './types/generation.types'
-export * from './types/template.types'
+export * from '../../../apps/server/src/modules/template/template.types'
 export * from './types/upload.types'
-export * from './types/notification.types'
-export * from './types/project.types'
-export * from './types/usage.types'
+export * from '../../../apps/server/src/modules/notification/notification.types'
+export * from '../../../apps/server/src/modules/project/project.types'
+export * from '../../../apps/server/src/modules/usage/usage.types'
 export * from './types/preview-manifest.types'
