+++ packages/shared/src/types/generation.types.ts
@@ -10,13 +10,12 @@ import {
 } from '../schemas/generation.schema'
 
 import type { PreviewManifestRef } from './preview-manifest.types'
-import type { CanvasDocument } from '../schemas/canvas.schema'
 
 export type GenerateWebsiteInput = {
     prompt: string
     userId: string
     projectId?: string
-    canvasState?: CanvasDocument
+    canvasState?: any
     model?: string
     onEvent?: (event: GenerationStreamEvent) => Promise<void> | void
 }
